+++ mangadex_downloader/auth/__init__.py
@@ -1,0 +1,2 @@
+from .oauth2 import OAuth2
+from .legacy import LegacyAuth